+++ src/linter/LintManager.ts
@@ -1,4 +1,4 @@
-import { Disposable, workspace, TextDocument, DiagnosticCollection, Diagnostic, languages, window, QuickPickItem, ProgressLocation } from "vscode";
+import { Disposable, workspace, TextDocument, window, QuickPickItem, ProgressLocation} from "vscode";
 
 import BaseLinter from "./BaseLinter";
 import IcarusLinter from "./IcarusLinter";
@@ -14,7 +14,7 @@ export default class LintManager {
 
     constructor() {
         workspace.onDidOpenTextDocument(this.lint, this, this.subscriptions);
-		workspace.onDidSaveTextDocument(this.lint, this, this.subscriptions);
+        workspace.onDidSaveTextDocument(this.lint, this, this.subscriptions);
         workspace.onDidCloseTextDocument(this.removeFileDiagnostics, this, this.subscriptions)
 
         workspace.onDidChangeConfiguration(this.configLinter, this, this.subscriptions);
@@ -53,8 +53,8 @@ export default class LintManager {
 
     lint(doc: TextDocument) {
         // Check for language id
-        let lang : string = window.activeTextEditor.document.languageId;
-        if(this.linter != null && window.activeTextEditor !== undefined && (lang === "verilog" || lang === "systemverilog"))
+        let lang : string = doc.languageId;
+        if(this.linter != null && (lang === "verilog" || lang === "systemverilog"))
             this.linter.startLint(doc);
     }