+++ CHANGELOG.md
@@ -3,9 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
-## [Unreleased] - 2019-01-29
+## [Unreleased]
 ### Changed
 - Replace fixed space with tab on snippets files
+- Properly fixed the bug which caused linter to run on all files
 
 ## [1.0.1] - 2019-01-01
 ### Fixed